Re-partition explorer tests
To get rid of `#![allow(dead_code)]` switch, functions specific to the
`explorer` example are moved into it.
